requirements.txt
amqp==2.6.1
anyjson==0.3.3
argon2-cffi==20.1.0
asgiref==3.3.4
backports.zoneinfo;python_version<"3.9"
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
cached-property==1.5.2
celery==4.3.0
cffi==1.14.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dj-database-url==0.5.0
Django==3.1.5
django-autoslug==1.9.8
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-crispy-forms==1.11.0
django-datetime-widget==0.9.3
django-formhelper==0.2
django-on-heroku==1.1.2
django-pandas==0.6.4
django-phonenumber-field==5.0.0
django-recaptcha==2.0.6
django-redis==4.12.1
django-render-block==0.8.1
django-resized==0.3.11
django-templated-email==2.3.0
django-utils-six==2.0
gunicorn==20.0.4
humanize==3.5.0
importlib-metadata==2.1.1
kombu==4.6.11
numpy==1.20.2
pandas==1.2.3
phonenumbers==8.12.18
Pillow==8.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.5
pytz==2020.5
redis==3.2.0
redis-server==6.0.9
simplejson==3.17.2
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
static3==0.7.0
tornado==6.1
typing-extensions==3.10.0.0
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.0
